@@ -3,13 +3,13 @@ namespace Serilog.Rendering;
 /// <summary>
 /// Class that provides reusable StringWriters to reduce memory allocations
 /// </summary>
-class ReusableStringWriter : StringWriter
+sealed class ReusableStringWriter : StringWriter
 {
     [ThreadStatic]
     static ReusableStringWriter? _pooledWriter;
 
     /// <summary>
-    /// Max capacity of StringBuilder we keep for next using
+    /// Max capacity of StringBuilder we keep for next using.
     /// </summary>
     internal const int StringBuilderCapacityThreshold = 32768;
 
@@ -52,8 +52,20 @@ class ReusableStringWriter : StringWriter
             return;
         }
         // We don't call base.Dispose because all it does is mark the writer as closed so it can't be
-        // written to and we want to keep it open as reusable writer.
+        // written to, and we want to keep it open as reusable writer.
         sb.Clear();
         _pooledWriter = this;
     }
+
+#if FEATURE_SPAN
+    public override void Write(ReadOnlySpan<char> buffer)
+    {
+        GetStringBuilder().Append(buffer);
+    }
+
+    public override void WriteLine(ReadOnlySpan<char> buffer)
+    {
+        GetStringBuilder().Append(buffer).AppendLine();
+    }
+#endif
 }
